backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0

# Web3 and Blockchain
web3==6.11.3
eth-account==0.9.0

# AI/ML - LangChain
langchain==0.1.0
langchain-openai==0.0.2
langchain-core==0.1.10
openai==1.3.0

# Utilities
python-dateutil==2.8.2

